Refactor of the CosmosDBSQLBulkExecutor
Focuses the functionality on it's ability to transform from a domain object into a Gremlin DB Object suitable for the Cosmos SQL SDK

Abstraction of the call to the container doesn't create much value and requires decisions to be made about what to do with the response. Decisions better suited for the client implementation.